+++ jobject-core/src/test/java/de/alpharogroup/reflection/ReflectionExtensionsTest.java
@@ -146,6 +146,30 @@ public class ReflectionExtensionsTest
 	}
 
 	/**
+	 * Test method for {@link ReflectionExtensions#getFieldValue(Object, String, Object)}.
+	 *
+	 * @throws NoSuchFieldException
+	 *             is thrown if no such field exists.
+	 * @throws SecurityException
+	 *             is thrown if a security manager says no.
+	 * @throws IllegalArgumentException
+	 *             is thrown if an illegal or inappropriate argument has been passed to a method.
+	 * @throws IllegalAccessException
+	 *             is thrown if an illegal on create an instance or access a method.
+	 */
+	@Test
+	public void testGetFieldValueObject() throws NoSuchFieldException, SecurityException,
+		IllegalArgumentException, IllegalAccessException
+	{
+		String expected;
+		String actual;
+		final Person person = Person.builder().name("Alex").build();
+		expected = "Alex";
+		actual = (String)ReflectionExtensions.getFieldValue(person, "name");
+		assertEquals(expected, actual);
+	}
+
+	/**
 	 * Test method for {@link ReflectionExtensions#getMethodNames(Class)}.
 	 */
 	@Test
@@ -225,6 +249,7 @@ public class ReflectionExtensionsTest
 		expected = new Person();
 		assertEquals(expected, actual);
 	}
+
 
 	/**
 	 * Test method for {@link ReflectionExtensions#newInstance(Object)}.