+++ CHANGELOG.md
@@ -8,6 +8,11 @@ ADDED:
 
 - new deploy scripts for module projects jobject-diff and jobject-evaluate
 
+CHANGED:
+
+- update of parent version to 4
+- update of test-objects dependency version from 4.26 to 4.28 
+- update of vintage-time dependency version from 4.11.0 to 4.12 
 
 Version 1.11
 -------------